disco
Managed to get buttons working so that each one changes color. Arranged in the possible arrangement for the game in a square matrix. Hopefully there is an easier way than hardcoding but nothing else worked for me. I hope that we can use vectors and inhereted classes in order to make the code simpler and better.